Fix nova_flavor to work with ansible 2.1.1.0
Without these fixes, int(ram) etc in create_flavor fail because they are None if they aren't specified.
And int(None) throws an exception: TypeError: int() argument must be a string or a number, not 'NoneType'.
Also the present_flavor.get_keys() is {} and extra_specs would be None, so inequal.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>